bug icon size default change to 21px (#151)
* bug icon size default change to 21px
The default styled size of the icons is 21px to make them small enough
to display properly in the button. However the SVG was setup with the
default size of 512px which causes a momentary fash of a large icon as
the page loads. This change sets the default size to 21px.

If a larger size is required in the future it can be overridden through
the styles prop just like the current implementation.

* fix(icons) - solid js icon default size

adjusted solid default icon size down to 21px to prevent flash of large icon prior to default styles render

* fix(icons) - update svelete icons default size 21px

prevent same issue as other libraries where large icon flashes prior to icon default styles being rendered

* Add changeset

---------

Co-authored-by: Andrew Smith <a.smith@silentworks.co.uk>